extract to useMarketplaceAdditionalSteps
